# storage/api-docs/api-documentation.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teachify API Documentation",
        "description": "API documentation for Teachify application - A comprehensive educational platform API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/{channel_slug}/groups/{group}/sessions": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "List sessions for a group",
                "operationId": "0541e3fe22effa0baf691edeebadeb92",
                "parameters": [
                    {
                        "name": "channel_slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "scheduled",
                                "live",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sessions"
                    },
                    "404": {
                        "description": "Group or channel not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Create a one-off or recurring session",
                "operationId": "5b0534ed08d8234a030c7182683980bc",
                "parameters": [
                    {
                        "name": "channel_slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "scheduled_at": {
                                        "description": "Required for one-off sessions",
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2026-06-10 09:00:00"
                                    },
                                    "duration_minutes": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "online",
                                            "offline"
                                        ]
                                    },
                                    "location": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "recurring_rule": {
                                        "properties": {
                                            "day": {
                                                "type": "string",
                                                "enum": [
                                                    "saturday",
                                                    "sunday",
                                                    "monday",
                                                    "tuesday",
                                                    "wednesday",
                                                    "thursday",
                                                    "friday"
                                                ]
                                            },
                                            "start_time": {
                                                "type": "string",
                                                "example": "09:00"
                                            },
                                            "end_time": {
                                                "type": "string",
                                                "example": "10:30",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Session created (or recurring sessions queued)"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Group or channel not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/{channel_slug}/groups/{group}/sessions/{session}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Get a session",
                "operationId": "f363ddbe6e95357ae0431e2ee0d0bb46",
                "parameters": [
                    {
                        "name": "channel_slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Cancel a session",
                "operationId": "02986dc8f08feec72f31ed980d7e6761",
                "parameters": [
                    {
                        "name": "channel_slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session cancelled"
                    },
                    "422": {
                        "description": "Cannot cancel a completed session"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Update a session (only if scheduled)",
                "operationId": "5d42d6ee6c5fc7c35c7faa62bd7ecec7",
                "parameters": [
                    {
                        "name": "channel_slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "datetime"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "scheduled",
                                            "live",
                                            "completed",
                                            "cancelled"
                                        ]
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session updated"
                    },
                    "422": {
                        "description": "Cannot edit non-scheduled session"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}